+++ test/src/template.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+class MyWidget extends StatelessWidget {
+  const MyWidget({
+    super.key,
+    required this.body,
+  });
+
+  final Widget body;
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      home: Scaffold(
+        appBar: AppBar(
+          title: Text('test template'),
+        ),
+        body: body,
+      ),
+    );
+  }
+}