+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # 0.1.0+1
 
 - feat: initial commit 🎉
+
+
+# 0.1.0+2
+
+- add date util from https://github.com/Sky24n/common_utils/blob/master/lib/src/date_util.dart
+- feat: date format supports zh_TW and en
+- add json_util, num_util
+- add Widgets.snackMessage